Minor Refactoring of Contract Package
