+++ MekHQ/src/mekhq/campaign/mission/contract/utilities/ActiveNegotiationMath.java
@@ -30,7 +30,7 @@
  * <https://www.xbox.com/en-US/developers/rules> and it is not endorsed by or
  * affiliated with Microsoft.
  */
-package mekhq.campaign.mission.contract.contractData;
+package mekhq.campaign.mission.contract.utilities;
 
 import mekhq.campaign.personnel.skills.enums.MarginOfSuccess;
