+++ MekHQ/src/mekhq/gui/dialog/markets/contractMarket/ContractNegotiationDialog.java
@@ -67,16 +67,16 @@ import mekhq.campaign.enums.DailyReportType;
 import mekhq.campaign.finances.Money;
 import mekhq.campaign.mission.contract.AbstractContract;
 import mekhq.campaign.mission.contract.contractData.ActiveNegotiationData;
-import mekhq.campaign.mission.contract.contractData.ActiveNegotiationMath;
 import mekhq.campaign.mission.contract.contractData.ChaosContractStepsTable;
 import mekhq.campaign.mission.contract.contractData.ContractTermsData;
 import mekhq.campaign.mission.contract.contractData.NegotiationData;
-import mekhq.campaign.mission.contract.contractData.NegotiationStepMath;
-import mekhq.campaign.mission.contract.contractData.NegotiationStepMath.Term;
 import mekhq.campaign.mission.contract.contractData.NonNegotiableTermsData;
 import mekhq.campaign.mission.contract.contractData.RentedFacilitiesData;
 import mekhq.campaign.mission.contract.contractGeneration.AbstractContractDeterminationPay;
 import mekhq.campaign.mission.contract.contractGeneration.negotiationsAndNPCs.TermFunding;
+import mekhq.campaign.mission.contract.utilities.ActiveNegotiationMath;
+import mekhq.campaign.mission.contract.utilities.NegotiationStepMath;
+import mekhq.campaign.mission.contract.utilities.NegotiationStepMath.Term;
 import mekhq.campaign.personnel.Person;
 import mekhq.campaign.personnel.PersonnelOptions;
 import mekhq.campaign.personnel.skills.ActionCheckResult;
